requirements.txt
python-dotenv==1.0.0
google-generativeai==0.8.5
Pillow==11.3.0